feat(directives): add @hasPermissions directive
The `@hasPermissions` directive, enforces the resolution of a field/type
to be guarded by a list of permissions. For more information on how to
use it, refer to:
https://github.com/profusion/apollo-validation-directives#haspermissions

As we are using a mocked environment, we need to apply the
SchemaDirectiveVisitor on top of the schema with the mocked resolvers -
since Apollo replace the resolvers with mocked ones. That's why we are
not using `mock: true` on ApolloServer creation but rather the explicit
function `addMockFunctionsToSchema`.

Change-Id: I30eb09b60866068a6693c1f0c919c8898f06b47e