feat(context): add JWT token to server context
If we receive a JWT token on the request, the server is going to
verify and decode it to get the incoming list of permissions by the
client. This permissions are attached to the server context so that
they can be validated by the `@hasPermissions` directive.

To decode the token, an cryptographic algorithm key must be given.
For now, this key is being read by one of the two places defined in
the .env file:
- PUBLIC_KEY_FILE: used for RSA asymmetric mechanisms, it is a file
path that contains the contents of the key (loaded dynamically by
the server)
- SECRET_KEY: used for HMAC symmetric mechanism, it is the value of
the key that can be generated using `uuidgen` for example.

Change-Id: Ib56ee3a0a299de46647f8efbdd2fd705cf0f75ab