chore(env): add .env example and documentation
The .env.example is a reference to .env configuration files used on
this project. This files are used to store environments configuration
that should be changed in different deploys. For now, we are using the
.env to store the keys used to decrypt the JWT token and set the
default port for GraphQL Server.

See README.md for more information on how to setup.

Change-Id: I34e116a93ffabf9d8feb77cd99b934db8419b771